feat: Added Caching to service worker tasks. style: Updated constant variables
The service worker now adds, deletes and returns cache based on 'activate', 'install' and 'fetch' listeners.
The static cache name is now a const combined from two other constants: the staticCacheName and staticCacheVer. The staticCacheName const is the name portion of the service worker, and the staticCacheVer is the version number of the service worker.

Changing the const variables this way helps later with deleting old cache, since the function checks if a new service worker has been registered based on the beginning name, not the version number.